refactor(workspace): add config and utils libs
