+++ libs/config/.babelrc
