+++ libs/config/README.md
@@ -1,0 +1,7 @@
+# config
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test config` to execute the unit tests via [Jest](https://jestjs.io).
