+++ libs/config/jest.config.ts
@@ -1,0 +1,13 @@
+/* eslint-disable */
+export default {
+    displayName: 'config',
+    preset: '../../jest.preset.js',
+    transform: {
+        '^.+\\.[tj]sx?$': [
+            '@swc/jest',
+            { jsc: { transform: { react: { runtime: 'automatic' } } } },
+        ],
+    },
+    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+    coverageDirectory: '../../coverage/libs/config',
+};
