+++ libs/config/src/constants.ts
@@ -1,0 +1,1 @@
+export const DOMAIN = 'https://shop.swiftsell.de';
