+++ libs/styles/.babelrc
@@ -1,6 +1,4 @@
 {
-  "presets": [
-    "@nrwl/next/babel"
-  ],
-  "plugins": []
+    "presets": ["@nrwl/next/babel"],
+    "plugins": []
 }
