+++ libs/utils/package.json
@@ -1,0 +1,4 @@
+{
+    "name": "@utils",
+    "version": "0.0.1"
+}
