+++ tsconfig.base.json
@@ -15,6 +15,7 @@
         "skipDefaultLibCheck": true,
         "baseUrl": ".",
         "paths": {
+            "@config": ["libs/config/src/index.ts"],
             "@hooks": ["libs/hooks/src/index.ts"],
             "@redux": ["libs/redux/src/index.ts"],
             "@shared": ["libs/shared/src/index.ts"],
